agents/1-tool_use/requirements.txt
jupyterlab
ipykernel
ipywidgets
openai-agents
openai>=1.12.0
python-dotenv>=1.0.0
youtube-transcript-api>=0.6.2